Included DBColumns class to db.py to call individual columns in the future.
Created a DBColumns(Enum) class with a class for each table inside followed by constants of each column in the table; this can be used in the future to call on individual columns when the time arises like with DBTables. Example: DBColumns.admins.GUILD_ID will return "guild_id" which is the name of the table.